requirements.txt
fastapi
uvicorn
pandas
scikit-learn
openai
python-dotenv
python-multipart